+++ ui/app/controllers/vault/cluster/clients.js
@@ -1,0 +1,3 @@
+import Controller from '@ember/controller';
+
+export default class ClientsController extends Controller {}
