+++ ui/app/routes/vault/cluster/clients.js
@@ -1,5 +1,6 @@
 import Route from '@ember/routing/route';
 import RSVP from 'rsvp';
+import { action } from '@ember/object';
 
 export default class ClientsRoute extends Route {
   async getVersionHistory() {
@@ -31,6 +32,16 @@ export default class ClientsRoute extends Route {
       config,
       // monthly: await this.store.queryRecord('clients/monthly', {}),
       versionHistory: this.getVersionHistory(),
+    });
+  }
+
+  @action
+  async loading(transition) {
+    // eslint-disable-next-line ember/no-controller-access-in-routes
+    let controller = this.controllerFor(this.routeName);
+    controller.set('currentlyLoading', true);
+    transition.promise.finally(function () {
+      controller.set('currentlyLoading', false);
     });
   }
 }
