+++ ui/app/routes/vault/cluster/clients/current.js
@@ -1,6 +1,5 @@
 import Route from '@ember/routing/route';
 import RSVP from 'rsvp';
-import { action } from '@ember/object';
 
 export default class CurrentRoute extends Route {
   async model() {
@@ -11,18 +10,5 @@ export default class CurrentRoute extends Route {
       monthly: await this.store.queryRecord('clients/monthly', {}),
       versionHistory: parentModel.versionHistory,
     });
-  }
-
-  @action
-  async loading(transition) {
-    // eslint-disable-next-line ember/no-controller-access-in-routes
-    let controller = this.controllerFor(this.routeName);
-    if (controller) {
-      // must use set here or it does not work see docs
-      controller.set('currentlyLoading', true);
-      transition.promise.finally(function () {
-        controller.set('currentlyLoading', false);
-      });
-    }
   }
 }
