+++ ui/app/templates/vault/cluster/clients/config.hbs
@@ -9,4 +9,4 @@
   </ToolbarActions>
 </Toolbar>
 
-<Clients::Config @model={{@model}} @isLoading={{this.currentlyLoading}} />
+<Clients::Config @model={{@model}} />
