+++ ui/app/templates/vault/cluster/clients/current.hbs
@@ -1,1 +1,1 @@
-<Clients::Current @model={{@model}} @isLoading={{this.currentlyLoading}} />
+<Clients::Current @model={{@model}} />
