+++ ui/app/templates/vault/cluster/clients/history.hbs
@@ -1,1 +1,1 @@
-<Clients::History @model={{@model}} @isLoading={{this.currentlyLoading}} />
+<Clients::History @model={{@model}} />
